workspace-backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
azure-storage-blob==12.22.0
azure-cosmos==4.6.0
pydantic==2.9.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.2
SQLAlchemy==2.0.35
